# plumbing: format/index perf, buffered reads, reflection removal