Large performance increase by buffering reads.

There were a few instances where binary.Read() would end up using reflection on &plumbing.Hash, rather than treating it as a byte slice. This has now been resolved.

On a 10MiB index file, buffering reduced the time it takes to read the index file and print out the filepaths from 3.9s to 0.7s. Removing the reflection overhead and `ReadUntil` usage reduced this to 0.07s. For comparison `git ls-files` took 0.08s :).